+++ changelog.md
@@ -1,6 +1,9 @@
 # Changelog
 
 ## Upcoming
+### Additions
+* Spawners now play the firecharge sound when spawning mobs
+
 ### Changes
 * Magic damage no longer bypasses invincibility frames
   * This should fix Pufferfishes instant kill